Integrate whisper thread settings
